feat: Add handleSignUp function to Login page
The commit adds a handleSignUp function to the Login page in the client/src/pages/Home/Login.jsx file. This function allows users to navigate to the SignUp page when they click on the "Sign Up" link. This change enhances the user experience by providing a seamless transition from the Login page to the SignUp page.