feat: Add isdeveloper field to User model
This commit adds the `isdeveloper` field to the User model in the `server/models/Users.model.js` file. The `isdeveloper` field is a boolean value that indicates whether a user is a developer or not. This change is necessary to distinguish between regular users and developers in the system and enable specific functionality for developers.