refactor: Update CreateDeveloperToken controller to store new token in DeveloperModel
This commit modifies the CreateDeveloperToken controller in the server/controllers/Developer.controller.js file. It updates the code to store the new token in the DeveloperModel, using the provided user and hashedToken. The previous implementation only logged the token and sent it as a response. This change improves the functionality by persisting the token in the database.

Refactor the CreateDeveloperToken controller to store new token in DeveloperModel