feat: Add DataGrids component for displaying tabular data
The code changes add a new component called DataGrids to the client/src/components directory. This component utilizes the @mui/x-data-grid library to render a tabular view of data. It includes columns for ID, first name, last name, age, and full name. The component also accepts an array of rows as a prop to populate the data. This addition enhances the application's ability to display and interact with tabular data.