feat: Update Users model to include role field
Update the Users model to include a role field. This field will store the role of the user, which can be one of the following: 'developer', 'hr', 'it', 'rb', or 'user'. The role field is required and has a default value of 'user'. This change enhances the functionality of the Users model and allows for better user management within the application.