```text Refactor OpenAccount component to handle account creation
Refactor the OpenAccount component in OpenAccount.jsx to handle the creation of a new account. Add a new function, handleCreateAccount, that sends a POST request to the server to create a new user and account. This enhancement improves the functionality of the application by allowing users to open new accounts.