feat: Update Customers component to handle active customer toggle
Refactor the Customers component in Customers.jsx to handle the active customer toggle. Add a new function, handleUpdateActiveCustomer, that sends a POST request to update the active status of a customer. This enhancement improves the functionality of the application by allowing the user to toggle the active status of a customer.